% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airDensity.R
\name{airDensity}
\alias{airDensity}
\title{Air density in \link{g cm^-3}}
\usage{
airDensity(
  Temp = 20,
  p = 1013.25,
  h = 50,
  units = c("deg.C", "hPa", "\%"),
  x_CO2 = 4e-04,
  opt = "CIMP2007"
)
}
\arguments{
\item{Temp}{temperature}
}
\value{
air density value
}
\description{
Calculates the density of the air in the laboratory according to two models that use
temperature, barometric pressure and relative humidity.
}
\details{
Jones1978 is the method A in the NIST doccument
CIMP2007
}
\examples{
airDensity(Temp = 22.3, p = 748.1, h = 37, units = c('deg.C', 'mmHg', '\%')) # [g/cm^3]
airDensity(Temp = 23.4, p = 612.3, h = 23, units = c('deg.C', 'mmHg', '\%')) # [g/cm^3]
airDensity(Temp = 23.4, p = 612.3, h = 23, units = c('deg.C', 'mmHg', '\%'), opt = 'A') # [g/cm^3]

}
