% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airDensityHASL.R
\name{airDensityHASL}
\alias{airDensityHASL}
\title{Air density in \link{g cm^-3} by using height above sea level information}
\usage{
airDensityHASL(HASL)
}
\arguments{
\item{HASL}{height altitude above sea level in meters}
}
\value{
air density value
}
\description{
Calculates the approximated density of the air in the laboratory according to its height above sea level
}
\examples{
airDensity(Temp = 22.3, p = 748.1, h = 37, units = c('deg.C', 'mmHg', '\%')) # [g/cm^3]
airDensity(Temp = 23.4, p = 612.3, h = 23, units = c('deg.C', 'mmHg', '\%')) # [g/cm^3]
airDensity(Temp = 23.4, p = 612.3, h = 23, units = c('deg.C', 'mmHg', '\%'), opt = 'A') # [g/cm^3]

}
