% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_convertMassUnitsSI.R
\name{convertMassUnitsSI}
\alias{convertMassUnitsSI}
\title{according to https://www.bipm.org/en/measurement-units/si-prefixes}
\usage{
convertMassUnitsSI(from, to, value)
}
\value{

}
\description{
according to https://www.bipm.org/en/measurement-units/si-prefixes
}
\examples{

}
