% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMassStandard.R
\name{createMassStandard}
\alias{createMassStandard}
\title{Creates a massStandard class object to be used in balance verification}
\usage{
createMassStandard(standardID, nominal)
}
\arguments{
\item{standardID}{the standard ID}

\item{nominal}{nominal mass of the mass standard}
}
\value{

}
\description{
Creates a massStandard class object to be used in balance verification
}
\examples{

}
