% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_massStandard.R
\docType{data}
\name{E2.MS.20g}
\alias{E2.MS.20g}
\title{Calibration data of a E2 class mass standard of nominal mass 20 g.}
\format{
A list with 25 objects of class \code{massStandard()}
}
\source{
Mass Laboratory - Instituto Nacional de Metrologia de Colombia.
CALIBRATION CERTIFICATE No. 4687.
}
\usage{
E2.MS.20g
}
\description{
An object of class \code{"massStandard"}, with calibration information
of a mass standards with nominal mass of 20 g,
performed by the Mass Laboratory at the Instituto Nacional de Metrologia de
Colombia (2020-08-12).
}
\keyword{datasets}
