% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MABC.R
\name{MABC}
\alias{MABC}
\title{Magnitude of the Air Buoyancy Correction}
\usage{
MABC(rho_s = 0.998, rho_w = 8, rho_air = airDensity())
}
\arguments{
\item{rho_s}{asragsd}
}
\description{
Calculates the magnitude
}
