% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_massStandard.R
\docType{data}
\name{massStandadsKit_E2}
\alias{massStandadsKit_E2}
\title{Calibration data of a weights kit in the interval 1 mg to 1 kg}
\format{
A list with 25 objects of class \code{massStandard()}
}
\source{
Mass Laboratory - Instituto Nacional de Metrologia de Colombia. CALIBRATION CERTIFICATE No. 4687.
}
\usage{
massStandadsKit_E2
}
\description{
A list with objects of class \code{massStandard()}, each with the
calibration information of a weight in the interval 1 mg to 1 kg,
performed by the Mass Laboratory at the Instituto Nacional de Metrologia de
Colombia (2020-08-12).
}
\keyword{datasets}
