% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimalCert}
\alias{minimalCert}
\title{Minimal \code{calibCert()} object}
\format{
An object of class \code{calibCert()}
}
\source{
Mass Laboratory - Instituto Nacional de Metrologia de Colombia. CALIBRATION CERTIFICATE No. 5143.
}
\usage{
minimalCert
}
\description{
An object of class \code{calibCert()} with the minimal calibration information of a hypothetical balance
calibrated using only two points.
}
\keyword{datasets}
