% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calibCert.R
\docType{data}
\name{MT.XP.2002}
\alias{MT.XP.2002}
\title{Calibration data for a balance Mettler Toledo XP 2002}
\format{
An object of class \code{calibCert()}
}
\source{
Mass Laboratory - Instituto Nacional de Metrologia de Colombia.
CALIBRATION CERTIFICATE No. 5142.
}
\usage{
MT.XP.2002
}
\description{
An object of class \code{calibCert()} with the calibration information of a balance
Mettler Toledo XP 2002, performed by the Mass Laboratory at the
Instituto Nacional de Metrologia de Colombia (2021-03-12).
}
\keyword{datasets}
