% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calibCert.R
\docType{data}
\name{MT.XPE.204}
\alias{MT.XPE.204}
\title{Calibration data for a balance Mettler Toledo XPE 204}
\format{
An object of class \code{calibCert()}
}
\source{
Mass Laboratory - Instituto Nacional de Metrologia de Colombia.
CALIBRATION CERTIFICATE No. 5143.
}
\usage{
MT.XPE.204
}
\description{
An object of class \code{calibCert()} with the calibration information of a balance
Mettler Toledo XPE 204, performed by the Mass Laboratory at the
Instituto Nacional de Metrologia de Colombia (2021-03-18).
}
\keyword{datasets}
