% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calibCert.R
\docType{data}
\name{MT.XPE.205}
\alias{MT.XPE.205}
\title{Calibration data for a balance Mettler Toledo XPE 205}
\format{
An object of class \code{calibCert()}
}
\source{
Mass Laboratory - Instituto Nacional de Metrologia de Colombia. CALIBRATION CERTIFICATE No. 4637.
}
\usage{
MT.XPE.205
}
\description{
An object of class \code{calibCert()} with the calibration information of a balance
Mettler Toledo XPE 205, performed by the Mass Laboratory at the Instituto Nacional de Metrologia de
Colombia (2020-07-14).
}
\keyword{datasets}
