% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizedError.R
\name{normalizedError}
\alias{normalizedError}
\title{Calculates normalized error in the conventional mass of a mass standard}
\usage{
normalizedError(
  reading,
  standard,
  calibCert = NULL,
  u_meas = NULL,
  u_massStandard = NULL
)
}
\arguments{
\item{reading}{balance reading for the standard mass.}

\item{standard}{object of class "massstandard" with the information of the
mass standard (see createMassStandard), or numeric value with the
conventional mass of the standard being used.}

\item{calibCert}{optional object of class "calibCert" with the calibration information}
}
\value{

}
\description{
Calculates normalized error in the conventional mass of a mass standard
}
\examples{

}
