% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.calibCert.R
\name{print.calibCert}
\alias{print.calibCert}
\title{S3 method for printing objects of class \code{"calibCert"}}
\usage{
\method{print}{calibCert}(x, complete = FALSE, nudeCertificate = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{"massStandard"}.}

\item{complete}{logical default to \code{FALSE}. If \code{TRUE}, all the
information contained in the object of class \code{calibCert} is shown.}

\item{nudeCertificate}{logical default to \code{FALSE}. If \code{TRUE},
the object of class \code{calibCert} is shown as a list.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
No return value, called for side effects.
}
\description{
The function prints objects of class \code{"calibCert"}.
}
\examples{
data(MT.XPE.204)
print(MT.XPE.204)
}
\seealso{
\code{\link[=calibCert]{calibCert()}}, \code{\link[=plot.calibCert]{plot.calibCert()}}
}
