% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.massStandard.R
\name{print.massStandard}
\alias{print.massStandard}
\title{S3 method for printing objects of class \code{"massStandard"}}
\usage{
\method{print}{massStandard}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"massStandard"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
The function prints objects of class \code{"massStandard"} in a certificate-like fashion.
}
\examples{
data(MT.XPE.204)
print(MT.XPE.204)
}
