% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertMABC.R
\name{uncertMABC}
\alias{uncertMABC}
\title{Standard uncertainty for the Air Buoyancy Correction Magnitude}
\usage{
uncertMABC(
  rho_s = 0.998,
  rho_w = 8,
  rho_air = NULL,
  u_rho_s = 1e-04,
  u_rho_w = 1e-04,
  u_rho_air = NULL,
  plot = FALSE
)
}
\description{
Standard uncertainty for the Air Buoyancy Correction Magnitude
}
