% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncert_masses.R
\name{uncertMassConv}
\alias{uncertMassConv}
\title{Uncertainty in conventional mass results}
\usage{
uncertMassConv(reading, units = NULL, calibCert, repValues = NULL, tare = TRUE)
}
\arguments{
\item{reading}{asdf}
}
\value{

}
\description{
Uncertainty in conventional mass results
}
\examples{

}
