% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibCert.R
\name{calibCert}
\alias{calibCert}
\title{creates a calibCert class object with the balance calibration information}
\usage{
calibCert(
  balanceID = "Balance",
  massSTD,
  indError,
  uncert,
  expanded = TRUE,
  k = 2,
  d,
  units = c("g", "mg", "mg", "mg"),
  classSTD = NULL,
  certSTD = NULL,
  barPress = NULL,
  ambTempe = NULL,
  relHumid = NULL,
  unitsENV = NULL,
  institution = NULL,
  date = NULL,
  details = NULL
)
}
\value{

}
\description{
creates a calibCert class object with the balance calibration information
}
\examples{
# MettlerToledo XPE204 2021-03-18. Certificate number 5143
massSTD <- c(0.01, 0.5, 1, 10, 20, 50, 100, 120, 150, 200, 220)  ## [g]
indError <- c(0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -0.2, -0.2) ## [mg]
uncert <- c(0.1, 0.1, 0.1, 0.1, 0.1, 0.2, 0.2, 0.3, 0.3, 0.4, 0.5) / 2 ## [mg]
d <- 0.1 ## [mg]
classSTD <- 'E2'
certSTD <- '1473 D-K 17296'
barPress <- c(750.4, 751.0) ## [hPa]
ambTemp <- c(17.4, 17.9) ## [deg.C]
relHumid <- c(70.5, 71.4) ## [\%]
unitsENV <- c('hPa', 'deg.C', '\%')
date <- '2021/03/18'
institution <- 'Instituto Nacional de Metrología de Colombia'

MT.XPE.204 <- calibCert(balanceID = 'MT XPE 204', massSTD = massSTD,
                        indError = indError, uncert = uncert, d = d,
                        units = c('g', 'mg', 'mg', 'mg'), classSTD = classSTD,
                        certSTD = certSTD, barPress = barPress, ambTemp = ambTemp,
                        relHumid = relHumid, unitsENV = unitsENV,
                        institution = institution)
}
