% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convMass.R
\name{convMass}
\alias{convMass}
\title{Converts balance reading to conventional mass}
\usage{
convMass(reading, units = NULL, calibCert)
}
\arguments{
\item{reading}{balance reading}
}
\value{

}
\description{
Conventional Mass: “The conventional value of the result of weighing a body in air is equal to the mass of a
standard, of conventionally chosen density, at a conventionally chosen temperature, which balances this body at this
reference temperature in air of conventionally chosen density.” The conventions are: reference density 8.0 g/cm3;
reference temperature 20 °C; normal air density 0.0012 g/cm3. See OIML D28 (2004)
}
\examples{

}
