% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massStandard.R
\name{massStandard}
\alias{massStandard}
\title{Creates an object of class \code{"massStandard"}.}
\usage{
massStandard(nominal, convMassCor, uncert, units = c("g", "mg", "mg"),
  serial = NULL, manufacturer = NULL, class = NULL, certificate = NULL,
  traceability = NULL, Temp = NULL, p = NULL, h = NULL,
  unitsENV = c("deg.C", "hPa", "\%"), expanded = TRUE, k = 2,
  rho = 8000, u_rho = 60, unitsrho = "kg/m^3", institution = NULL,
  date = NULL, add.info = NULL)
}
\arguments{
\item{nominal}{Nominal mass of the mass standard.}

\item{convMassCor}{Conventional mass correction for the mass standard.}

\item{uncert}{Standard uncertainty of the conventional mass.}

\item{units}{Character vector of length 3 with the units of \code{nominal}, \code{convMass} and \code{uncert}, respectively.
Default is \code{c('g', 'mg', 'mg')}.}

\item{serial}{Serial number of the mass standard.}

\item{manufacturer}{Character with the manufacturer of the weights.}

\item{class}{Character with the claimed class of the mass standard according to REF REF (XXXX).}

\item{certificate}{Character with the calibration certificate.}

\item{traceability}{Character with information regarding the traceability of the calibration.}

\item{Temp}{Ambient temperature at the moment of the calibration.}

\item{p}{Barometric pressure at the moment of the calibration.}

\item{h}{Relative humidity at the moment of the calibration.}

\item{unitsENV}{Character vector of length 3 with the units of \code{p}, \code{Temp} and \code{h}.
Default is \code{c('deg.C', 'hPa', '\%')}. See \strong{unitsENV} below for more options.}

\item{expanded}{If \code{TRUE} (the default), \code{uncert} is assumed to be expanded uncertainties instead of standard uncertainties.}

\item{k}{Coverage factor for \code{uncert} when \code{expanded = TRUE}.}

\item{rho}{Density...}

\item{u_rho}{Uncertainty in ...}

\item{unitsrho}{Units}

\item{institution}{Character with the institution that performed the calibration.}

\item{date}{Character with the date of the calibration.}

\item{add.info}{Named list or vector with any additional details included in calibration certificate.}
}
\value{
Object of class \code{"massStandard"}
}
\description{
The can object of class \code{"massStandard"} be used in routine balance verification to calculate
normalized error. See \code{\link[=normalizedError]{normalizedError()}}.
}
\details{
Several objects of class \code{"massStandard"} can be concatenated in a single list if the mass
standards belongs to the same
}
\examples{
Box1.E2.10g <- createMassStandard(nominal = 10, convMass = 10.001,
                                  uncert = 0.1, units = c('g', 'mg', 'mg'))
print(Box1.E2.10g)
}
\seealso{
\code{\link[=normalizedError]{normalizedError()}}
}
