% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massStandard.R
\name{massStandard}
\alias{massStandard}
\title{Creates an object of class \code{"massStandard"}.}
\usage{
massStandard(nominal, convMassCor, uncert, units = c("g", "mg", "mg"),
  serial = NULL, manufacturer = NULL, class = NULL, certificate = NULL,
  traceability = NULL, Temp = NULL, p = NULL, h = NULL,
  unitsENV = c("deg.C", "hPa", "\%"), expanded = TRUE, k = 2,
  rho = 8000, u_rho = 60, unitsrho = "kg/m^3", institution = NULL,
  date = NULL, add.info = NULL, partofakit = FALSE)
}
\arguments{
\item{nominal}{Nominal mass of the mass standard.}

\item{convMassCor}{Conventional mass correction for the mass standard.}

\item{uncert}{Standard uncertainty of the conventional mass correction.}

\item{units}{Character vector of length 3 with the units of \code{nominal}, \code{convMass} and
\code{uncert}, respectively.
Default is \code{c('g', 'mg', 'mg')}.}

\item{serial}{Serial number of the mass standard or mass standards kit.}

\item{manufacturer}{Character with the manufacturer of the mass standard or mass standards kit.}

\item{class}{Character with the claimed class of the mass standard or mass standards kit,
according to REF REF (XXXX).}

\item{certificate}{Character with the calibration certificate.}

\item{traceability}{Character with information regarding the traceability of the calibration.}

\item{Temp}{Ambient temperature at the moment of the calibration.}

\item{p}{Barometric pressure at the moment of the calibration.}

\item{h}{Relative humidity at the moment of the calibration.}

\item{unitsENV}{Character vector of length 3 with the units of \code{p}, \code{Temp} and \code{h}.
Default is \code{c('deg.C', 'hPa', '\%')}. See \strong{unitsENV} below for more options.}

\item{expanded}{If \code{TRUE} (the default), \code{uncert} is assumed to be expanded uncertainties instead of standard uncertainties.}

\item{k}{Coverage factor for \code{uncert} when \code{expanded = TRUE}.}

\item{rho}{Density of the mass standard.}

\item{u_rho}{Uncertainty in the density of the mass standard.}

\item{unitsrho}{Units of the density of the mass standard. Default is \code{'kg/m^3'}.}

\item{institution}{Character with the institution that performed the calibration.}

\item{date}{Character with the date of the calibration.}

\item{add.info}{Named list or vector with any additional details included in calibration certificate.}

\item{partofakit}{Logical. Is the mass standard part of kit?}
}
\value{
Object of class \code{"massStandard"} with the information of a calibrated mass standard.
}
\description{
The object of class \code{"massStandard"} contains the calibration information of a mass standard
that is used in routine balance verification (e g. to calculate
normalized error. See \code{\link[=normalizedError]{normalizedError()}}). A version to store information about several mass
standards that belong to the same kit is \code{\link[=massStandardKit]{massStandardKit()}}.
}
\section{unitsENV}{

Pressure units (\code{p}) can be any of \code{'mmHg'}, \code{'Pa'}, \code{'hPa'} or \code{'kPa'}.
Temperature units (\code{Temp}) can be either \code{'deg.C'} (for Celsius degrees) or \code{'K'}.
Relative humidity (\code{h}) can be expressed as fraction (\code{'frac'}) or as percentage (\code{'\%'}).
A typical arrangement for the parameter \code{unitsENV} could be \code{c('hPa', 'deg.C', '\%')}.
}

\examples{
singleMS.E2.10g <- massStandard(nominal = 10, convMassCor = 0.001, uncert = 0.1,
                                units = c('g', 'mg', 'mg'))
print(singleMS.E2.10g)
}
\seealso{
\code{\link[=massStandardKit]{massStandardKit()}}, \code{\link[=normalizedError]{normalizedError()}}
}
