% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.calibCert.R
\name{print.calibCert}
\alias{print.calibCert}
\title{S3 method for printing objects of class calibCert}
\usage{
\method{print}{calibCert}(x)
}
\description{
S3 method for printing objects of class calibCert
}
\examples{

}
