% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertAirDensity.R
\name{uncertAirDensity}
\alias{uncertAirDensity}
\title{Uncertainties in air density calculation}
\usage{
uncertAirDensity(
  value = NULL,
  opt = "CIMP2007",
  Temp = 20,
  p = 1013.25,
  h = 50,
  u_Temp = 2.9,
  u_p = 10.1,
  u_h = 11.3,
  plot = FALSE,
  units = c("deg.C", "hPa", "\%"),
  printRelSD = TRUE
)
}
\arguments{
\item{HASL}{height altitude above sea level in meters}
}
\value{
air density value
}
\description{
Calculates the approximated density of the air in the laboratory according to its height above sea level
}
\examples{
airDensity(Temp = 22.3, p = 748.1, h = 37, units = c('deg.C', 'mmHg', '\%')) # [g/cm^3]
airDensity(Temp = 23.4, p = 612.3, h = 23, units = c('deg.C', 'mmHg', '\%')) # [g/cm^3]
airDensity(Temp = 23.4, p = 612.3, h = 23, units = c('deg.C', 'mmHg', '\%'), opt = 'A') # [g/cm^3]

}
