% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncert_masses.R
\name{uncertErrorCorr}
\alias{uncertErrorCorr}
\title{Uncertainty associated to error correction}
\usage{
uncertErrorCorr(reading, units = NULL, calibCert)
}
\arguments{
\item{reading}{asdf}
}
\value{

}
\description{
Uncertainty associated to error correction
}
\examples{

}
