% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncert_masses.R
\name{uncertReading}
\alias{uncertReading}
\title{Uncertainty in balance reading value}
\usage{
uncertReading(calibCert, repValues = NULL, units = NULL)
}
\arguments{
\item{reading}{asdf}
}
\value{

}
\description{
Uncertainty in balance reading value
}
\examples{

}
