% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncert_masses.R
\name{uncertReading}
\alias{uncertReading}
\title{Uncertainty of balance readings}
\usage{
uncertReading(calibCert, repValues = NULL, d = NULL, units = NULL)
}
\arguments{
\item{calibCert}{Object of class \code{"calibCert"} with the calibration data of the balance.
See \code{\link[=calibCert]{calibCert()}} for details.}

\item{repValues}{Numeric vector with balance readings for the same mass standard under repeatability conditions.}

\item{d}{Scale division used in balance reading. Useful when operating the balance at a
division scale different from that specified in the calibration certificate (e.g. when operating a
semimicro balance with only four digital places). If not provided
the one in the calibration certificate is used.}

\item{units}{Character with the units of \code{reading}. Must be a SI unit. If not provided, the
balance standard units are assumed. See \code{\link[=calibCert]{calibCert()}} for details.}
}
\value{
Uncertainty of balance readings
}
\description{
Combination of readability uncertainty due to scale division and lack of repeatability
}
\examples{
data(minimalCert)
uncertReading(calibCert = minimalCert,
              repValues = c(5.0000, 5.0000, 4.9999, 4.9999, 4.9999,
                            4.9999, 4.9999, 4.9999, 4.9999, 4.9999))
}
\seealso{
\code{\link[=uncertErrorCorr]{uncertErrorCorr()}}, \code{\link[=uncertMassConv]{uncertMassConv()}}
}
